+++ src/components/LoginComponent.jsx
@@ -1,0 +1,80 @@
+import React, { useState } from "react";
+import { Target } from "lucide-react";
+
+const LoginComponent = ({ users, setCurrentUser, setCurrentView }) => {
+  const [credentials, setCredentials] = useState({ username: '', password: '' });
+  const [error, setError] = useState('');
+
+  const handleLogin = (e) => {
+    e.preventDefault();
+    const user = users.find(
+      (u) =>
+        u.username === credentials.username &&
+        u.password === credentials.password &&
+        u.status === 'activo'
+    );
+
+    if (user) {
+      setCurrentUser(user);
+      setCurrentView('dashboard');
+      setError('');
+    } else {
+      setError('Credenciales inválidas o usuario inactivo');
+    }
+  };
+
+  return (
+    <div className="min-h-screen bg-gradient-to-br from-blue-600 to-purple-700 flex items-center justify-center p-4">
+      <div className="bg-white rounded-2xl shadow-2xl p-8 w-full max-w-md">
+        <div className="text-center mb-8">
+          <div className="bg-blue-100 w-16 h-16 rounded-full flex items-center justify-center mx-auto mb-4">
+            <Target className="text-blue-600" size={32} />
+          </div>
+          <h1 className="text-3xl font-bold text-gray-800 mb-2">Sistema de Proyectos</h1>
+          <p className="text-gray-600">Ingresa tus credenciales para continuar</p>
+        </div>
+
+        <form onSubmit={handleLogin} className="space-y-6">
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-2">Usuario</label>
+            <input
+              type="text"
+              required
+              value={credentials.username}
+              onChange={(e) => setCredentials({ ...credentials, username: e.target.value })}
+              className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+              placeholder="Ingresa tu usuario"
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700 mb-2">Contraseña</label>
+            <input
+              type="password"
+              required
+              value={credentials.password}
+              onChange={(e) => setCredentials({ ...credentials, password: e.target.value })}
+              className="w-full px-4 py-3 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500 focus:border-transparent"
+              placeholder="Ingresa tu contraseña"
+            />
+          </div>
+
+          {error && (
+            <div className="bg-red-50 border border-red-200 text-red-700 px-4 py-3 rounded-lg text-sm">
+              {error}
+            </div>
+          )}
+
+          <button
+            type="submit"
+            className="w-full bg-blue-600 hover:bg-blue-700 text-white py-3 rounded-lg font-medium transition-colors"
+          >
+            Iniciar Sesión
+          </button>
+        </form>
+      </div>
+    </div>
+  );
+};
+
+export default LoginComponent;
